Phase 8: AMD64 シグナル配信を実装 (kill + handler 往復)
新規テスト 1 本 + 既存 sys_kill64 を SKIP 解除:

  sys_signal_delivery64  rt_sigaction で SIGUSR1 ハンドラを登録 →
                         self-kill → ハンドラ実行 → main へ復帰を確認
  sys_kill64             非存在 pid に kill して -ESRCH (-3) を確認

実装内容:

1. Cpu64.eval ループにシグナルチェックを追加
   各命令実行前に process.psig() を見て pending があれば
   SIG_IGN/SIG_DFL/カスタムハンドラ を分岐処理。カスタムハンドラは
   "rip を push してハンドラへジャンプ + RDI = signum" の単純実装。
   ハンドラの ret で push した rip を pop して復帰する。

2. amd64_kill (syscall #62)
   target pid を kernel.find_process で探し、見つかれば recv(sig)。
   見つからなければ -ESRCH (-3) を返す。pid<=0 は self へ送信。

3. amd64_rt_sigaction (syscall #13)
   sigaction 構造体の sa_handler (offset 0, 8 バイト) を読み、
   process.set_sigaction(signum, handler) で登録。oldact が non-null
   なら旧ハンドラを書き戻す。

4. amd64_rt_sigreturn (syscall #15)
   我々の単純実装では rip push + ret で復帰するため通常使われない。
   glibc の sa_restorer 経路と互換にするためのスタブを用意。

5. Kernel.find_process(pid) を追加
   ptable から指定 pid のプロセスを検索する。

6. Siginfo / Signal クラスの func_adrs を int → long に拡張
   x86-64 の 64-bit ハンドラアドレスを保持するため。i386 路は
   Process.run の signal 処理で (int)func_adrs にキャストして対応。

回帰テスト: 42 本中 41 PASS / 0 FAIL / 1 SKIP (sys_chmod64 のみ)。

+++ src/main/java/emulin/Process.java
@@ -202,7 +202,7 @@ public class Process extends Signal {
     }
     else {               // それ以外のプロセス
       int sig;
-      int func_adrs;
+      long func_adrs;
       // CPUの実行サイクルに入る
       while( !exit_flag ) {
 	if( exit_flag ) {
@@ -247,7 +247,7 @@ public class Process extends Signal {
 		if( !done ) {
 		    // func_adrs で指し示す関数を実行する。
 		    mem.store32( sig_no_embed_adrs, sig );
-		    mem.store32( handler_embed_adrs, func_adrs );
+		    mem.store32( handler_embed_adrs, (int)func_adrs );
 		    cpu.set_signal_handler( ip, handler_hook );
 		    ip = cpu.get_ip( );
 		    done = true;
